+++ src/session.jl
@@ -840,6 +840,8 @@ It can return any of:
 - `session`: The [`Session`](@ref) to authenticate.
 - `password=nothing`: A password to authenticate with. Pass this if
   `authenticate()` previously returned `AuthMethod_Password`.
+- `privkey=nothing`: A `SshKey` to authenticate with. Pass
+  this if `authenticate()` previously returned `AuthMethod_PublicKey`.
 - `kbdint_answers=nothing`: Answers to keyboard-interactive prompts from the
   server. Use [`userauth_kbdint_getprompts()`](@ref) to get the prompts if
   `authenticate()` returns `AuthMethod_Interactive` and then pass the answers in
@@ -854,8 +856,7 @@ It can return any of:
   available.
 - `LibSSHException`: If there's an internal error and `throw=true`.
 """
-function authenticate(session::Session; password=nothing, kbdint_answers=nothing,
-                      throw=true)
+function authenticate(session::Session; password=nothing, privkey::Union{PKI.SshKey, Nothing}=nothing, kbdint_answers=nothing, throw=true)
     if !isconnected(session)
         Base.throw(ArgumentError("Session is disconnected, cannot authenticate"))
     elseif !isnothing(password) && !isnothing(kbdint_answers)
@@ -876,9 +877,11 @@ function authenticate(session::Session; password=nothing, kbdint_answers=nothing
 
     # First we check if any of the input arguments have been passed, and we
     # attempt authentication if so.
-    if !isnothing(password) || !isnothing(kbdint_answers)
+    if !isnothing(password) || !isnothing(kbdint_answers) || !isnothing(privkey)
         status = if !isnothing(password)
             userauth_password(session, password; throw)
+        elseif !isnothing(privkey)
+            userauth_publickey(session, privkey; throw)
         else
             _try_userauth_kbdint(session, kbdint_answers, throw)
         end
@@ -927,6 +930,11 @@ function authenticate(session::Session; password=nothing, kbdint_answers=nothing
     # Then password auth
     if _can_attempt_auth(session, AuthMethod_Password)
         return AuthMethod_Password
+    end
+
+    # Then public key auth
+    if _can_attempt_auth(session, AuthMethod_PublicKey)
+        return AuthMethod_PublicKey
     end
 
     # Then keyboard-interactive auth
@@ -1138,6 +1146,65 @@ function userauth_password(session::Session, password::String; throw=true)
             return ret
         end
     end
+end
+
+"""
+$(TYPEDSIGNATURES)
+
+Authenticate by username and private key. The username will be taken from
+`session.user`.
+
+# Arguments
+- `session`: The session to authenticate.
+- `path`: The private key file path to authenticate with.
+- `passphrase=nothing`: An optional passphrase for the private key, if it's encrypted.
+- `throw=true`: Whether to throw if there's an internal error while
+  authenticating (`AuthStatus_Error`).
+
+# Throws
+- `ArgumentError`: If the session isn't connected.
+- `LibSSHException`: If there was an internal error, unless `throw=false`.
+
+Wrapper around [`lib.ssh_userauth_publickey()`](@ref).
+"""
+function userauth_publickey(session::Session, path::AbstractString; passphrase=nothing, throw=true)
+    privkey = PKI.import_privkey_file(path; passphrase)
+    return userauth_publickey(session, privkey; throw)
+end
+
+"""
+$(TYPEDSIGNATURES)
+
+Authenticate by username and private key. The username will be taken from
+`session.user`.
+
+# Arguments
+- `session`: The session to authenticate.
+- `privkey`: The private key to authenticate with, as a `SshKey` object.
+- `throw=true`: Whether to throw if there's an internal error while
+  authenticating (`AuthStatus_Error`).
+
+# Throws
+- `ArgumentError`: If the session isn't connected.
+- `LibSSHException`: If there was an internal error, unless `throw=false`.
+
+Wrapper around [`lib.ssh_userauth_publickey()`](@ref).
+"""
+function userauth_publickey(session::Session, privkey::PKI.SshKey; throw=true)
+    if !isconnected(session)
+        Base.throw(ArgumentError("Session is disconnected, cannot authenticate until it's connected"))
+    end
+
+    ret = _session_trywait(session) do
+        LibSSH.lib.ssh_userauth_publickey(session, C_NULL, privkey.ptr)
+    end
+    status = AuthStatus(ret)
+
+    if status == AuthStatus_Error && throw
+        Base.throw(LibSSHException("Got AuthStatus_Error (SSH_AUTH_ERROR) when authenticating"))
+    end
+
+    return status
 end
 
 """